fix(relpro): set 'version' parameter
Tasks like `mark-as-shipped` expect a version parameter to exist. This
is supposed to be set by the release-promotion action, but I neglected
to add this.

Typically we just assume that there's a file called `version.txt` at the
root of the repo, but this PR implements a new `version-parser` key in
config.yml. This defaults to a parser that reads `version.txt`, but can
be used to point to a custom function in case we want to support storing
the version in other locations.

This function takes the parameters as input so that it can make
decisions around what version to return. For example, in
`mozilla-vpn-client`, there are two distinct products that can be
shipped (client and addons). This will allow them to determine a version
based on things like the `shipping-phase`.