feat: use 'scriptworker.scope-prefix' config for mark-as-shipped
Currently we look at a `shipit.scope-prefix` config to get the scope
prefix. However, `scriptworker.scope-prefix` is already a standard.

This implements the ability to use that config in a backwards compatible
manner. I.e, it will first check the `shipit` config, then the
`scriptworker` one if none was found.